Add tests for form fields and tagged parameters
