+++ akka-http/src/test/scala/pl/iterators/kebs/matchers/AkkaHttpMatchersTests.scala
@@ -1,36 +1,14 @@
+package pl.iterators.kebs.matchers
+
 import akka.http.scaladsl.server.Directives
 import akka.http.scaladsl.testkit.ScalatestRouteTest
-import enumeratum.{Enum, EnumEntry}
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
+import pl.iterators.kebs.Domain._
 import pl.iterators.kebs.instances.TimeInstances.{DayOfWeekInt, InstantEpochMilliLong, ZonedDateTimeString}
-import pl.iterators.kebs.matchers._
-import pl.iterators.kebs.tag.meta.tagged
-import pl.iterators.kebs.tagged._
 
 import java.time.{DayOfWeek, Instant, ZonedDateTime}
-import java.util.UUID
-
-@tagged trait Tags {
-  trait IdTag
-  type Id = Long @@ IdTag
-
-  trait TestIdTag
-  type TestId = UUIDId @@ TestIdTag
-
-  trait TestDoubleTag
-  type TestDouble = Double @@ TestDoubleTag
-
-  type UUIDId = UUID
-  object UUIDId {
-    def generate[T]: UUIDId @@ T = UUID.randomUUID().taggedWith[T]
-    def fromString[T](str: String): UUIDId @@ T =
-      UUID.fromString(str).taggedWith[T]
-  }
-}
-
-object Domain extends Tags
 
 class AkkaHttpMatchersTests
     extends AnyFunSuite
@@ -68,8 +46,6 @@ class AkkaHttpMatchersTests
       responseAs[String] shouldEqual "1621258399"
     }
   }
-
-  import Domain._
   test("Extract Double instance ") {
     val testRoute = path("test" / DoubleNumber.as[TestDouble]) { test =>
       complete(test.toString)
@@ -95,16 +71,6 @@ class AkkaHttpMatchersTests
     Get("/test/ce7a7cf1-8c00-49a9-a963-9fd119dd0642") ~> testRoute ~> check {
       responseAs[String] shouldEqual "ce7a7cf1-8c00-49a9-a963-9fd119dd0642"
     }
-  }
-
-  sealed trait Greeting extends EnumEntry
-  object Greeting extends Enum[Greeting] {
-    case object Hello   extends Greeting
-    case object GoodBye extends Greeting
-    case object Hi      extends Greeting
-    case object Bye     extends Greeting
-
-    val values = findValues
   }
 
   test("Extract Enum instance") {
